requirements.txt
fastapi
uvicorn
pillow
numpy
tensorflow
python-multipart
gdown
